requirements.txt
# LLM e Processamento de Linguagem Natural
langchain>=0.0.350
pydantic>=2.0
pydantic-settings>=2.0
openai>=1.3.7
cohere>=4.37
google-generativeai>=0.3.1
ollama>=0.1.5
transformers>=4.35.2
torch>=2.1.1

# Framework Web e Ferramentas
fastapi>=0.105.0
uvicorn>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.2
email-validator>=2.1.0

# Base de Dados e ORM
supabase>=2.8.1
sqlalchemy>=2.0.23
alembic>=1.13.1
redis>=5.0.1

# Utilitários
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
logfire>=3.1.0

# CLI
click>=8.0.0
rich>=10.0.0
aiohttp>=3.8.0

# Testes
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov==4.1.0
locust>=2.32.6
psutil>=5.9.0

# Prometheus
prometheus-client>=0.17.0

# Prometheus FastAPI Instrumentator
prometheus-fastapi-instrumentator>=6.1.0 

structlog>=23.1.0
